Pull out the traceroute engine into a separate class
