+++ projects/batfish/src/main/java/org/batfish/dataplane/ibdp/IncrementalDataPlane.java
@@ -21,7 +21,6 @@ import org.batfish.datamodel.Fib;
 import org.batfish.datamodel.GenericRib;
 import org.batfish.datamodel.Ip;
 import org.batfish.datamodel.Topology;
-import org.batfish.datamodel.collections.NodeInterfacePair;
 
 public class IncrementalDataPlane implements Serializable, DataPlane {
 
@@ -29,15 +28,13 @@ public class IncrementalDataPlane implements Serializable, DataPlane {
 
   private transient Network<BgpNeighbor, BgpSession> _bgpTopology;
 
-  Set<NodeInterfacePair> _flowSinks;
-
-  Map<Ip, Set<String>> _ipOwners;
+  private Map<Ip, Set<String>> _ipOwners;
 
   private Map<Ip, String> _ipOwnersSimple;
 
   Map<String, Node> _nodes;
 
-  Topology _topology;
+  private Topology _topology;
 
   @Override
   public Map<String, Map<String, Fib>> getFibs() {
@@ -58,6 +55,7 @@ public class IncrementalDataPlane implements Serializable, DataPlane {
                                 Entry::getKey, eVr -> eVr.getValue()._fib))));
   }
 
+  @Override
   public Map<Ip, Set<String>> getIpOwners() {
     return _ipOwners;
   }
@@ -89,6 +87,11 @@ public class IncrementalDataPlane implements Serializable, DataPlane {
   }
 
   @Override
+  public Topology getTopology() {
+    return _topology;
+  }
+
+  @Override
   public SortedSet<Edge> getTopologyEdges() {
     return _topology.getEdges();
   }
@@ -99,10 +102,6 @@ public class IncrementalDataPlane implements Serializable, DataPlane {
       Map<Ip, String> ipOwnersSimple) {
     setIpOwners(ipOwners);
     setIpOwnersSimple(ipOwnersSimple);
-  }
-
-  public void setFlowSinks(Set<NodeInterfacePair> flowSinks) {
-    _flowSinks = flowSinks;
   }
 
   public void setIpOwners(Map<Ip, Set<String>> ipOwners) {
@@ -124,6 +123,14 @@ public class IncrementalDataPlane implements Serializable, DataPlane {
   @Override
   public Network<BgpNeighbor, BgpSession> getBgpTopology() {
     return _bgpTopology;
+  }
+
+  @Override
+  public Map<String, Configuration> getConfigurations() {
+    return _nodes
+        .entrySet()
+        .stream()
+        .collect(ImmutableMap.toImmutableMap(Entry::getKey, e -> e.getValue().getConfiguration()));
   }
 
   void setBgpTopology(Network<BgpNeighbor, BgpSession> bgpTopology) {
