+++ projects/batfish/src/main/java/org/batfish/dataplane/ibdp/IncrementalDataPlanePlugin.java
@@ -22,6 +22,7 @@ import org.batfish.datamodel.Topology;
 import org.batfish.datamodel.answers.Answer;
 import org.batfish.datamodel.answers.BdpAnswerElement;
 import org.batfish.datamodel.collections.IbgpTopology;
+import org.batfish.dataplane.TracerouteEngineImpl;
 
 @AutoService(Plugin.class)
 public class IncrementalDataPlanePlugin extends DataPlanePlugin {
@@ -128,7 +129,8 @@ public class IncrementalDataPlanePlugin extends DataPlanePlugin {
   public void processFlows(Set<Flow> flows, DataPlane dataPlane, boolean ignoreAcls) {
     _flowTraces.put(
         (IncrementalDataPlane) dataPlane,
-        _engine.processFlows((IncrementalDataPlane) dataPlane, flows, ignoreAcls));
+        TracerouteEngineImpl.getInstance()
+            .processFlows(dataPlane, flows, dataPlane.getFibs(), ignoreAcls));
   }
 
   private IncrementalDataPlane loadDataPlane() {
