+++ projects/batfish/src/test/java/org/batfish/dataplane/bdp/BdpDataPlanePluginTest.java
@@ -1,6 +1,5 @@
 package org.batfish.dataplane.bdp;
 
-import static java.util.Collections.singletonList;
 import static org.batfish.datamodel.Configuration.DEFAULT_VRF_NAME;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
@@ -19,7 +18,6 @@ import static org.junit.Assert.assertTrue;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSortedMap;
-import com.google.common.collect.Lists;
 import com.google.common.graph.Network;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -36,7 +34,6 @@ import java.util.TreeMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.batfish.common.BatfishException;
 import org.batfish.common.BatfishLogger;
 import org.batfish.common.BdpOscillationException;
 import org.batfish.common.plugin.DataPlanePlugin;
@@ -50,24 +47,18 @@ import org.batfish.datamodel.BgpSession;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.ConfigurationFormat;
 import org.batfish.datamodel.DataPlane;
-import org.batfish.datamodel.Flow;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.InterfaceAddress;
 import org.batfish.datamodel.Ip;
-import org.batfish.datamodel.IpAccessList;
-import org.batfish.datamodel.IpAccessListLine;
-import org.batfish.datamodel.LineAction;
 import org.batfish.datamodel.MultipathEquivalentAsPathMatchMode;
 import org.batfish.datamodel.NetworkFactory;
 import org.batfish.datamodel.OriginType;
 import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.Route;
 import org.batfish.datamodel.RoutingProtocol;
-import org.batfish.datamodel.SourceNat;
 import org.batfish.datamodel.StaticRoute;
 import org.batfish.datamodel.Topology;
 import org.batfish.datamodel.Vrf;
-import org.batfish.datamodel.acl.TrueExpr;
 import org.batfish.datamodel.answers.BdpAnswerElement;
 import org.batfish.datamodel.collections.RoutesByVrf;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
@@ -209,96 +200,6 @@ public class BdpDataPlanePluginTest {
 
     // Test that compute Data Plane finishes in a finite time
     dataPlanePlugin.computeDataPlane(false);
-  }
-
-  private static Flow makeFlow() {
-    Flow.Builder builder = new Flow.Builder();
-    builder.setSrcIp(new Ip("1.2.3.4"));
-    builder.setIngressNode("foo");
-    builder.setTag("TEST");
-    return builder.build();
-  }
-
-  private static IpAccessList makeAcl(String name, LineAction action) {
-    IpAccessListLine aclLine =
-        IpAccessListLine.builder().setAction(action).setMatchCondition(TrueExpr.INSTANCE).build();
-    return new IpAccessList(name, singletonList(aclLine));
-  }
-
-  @Test
-  public void testApplySourceNatSingleAclMatch() {
-    Flow flow = makeFlow();
-
-    SourceNat nat = new SourceNat();
-    nat.setAcl(makeAcl("accept", LineAction.ACCEPT));
-    nat.setPoolIpFirst(new Ip("4.5.6.7"));
-
-    Flow transformed =
-        BdpEngine.applySourceNat(
-            flow, null, ImmutableMap.of(), ImmutableMap.of(), singletonList(nat));
-    assertThat(transformed.getSrcIp(), equalTo(new Ip("4.5.6.7")));
-  }
-
-  @Test
-  public void testApplySourceNatSingleAclNoMatch() {
-    Flow flow = makeFlow();
-
-    SourceNat nat = new SourceNat();
-    nat.setAcl(makeAcl("reject", LineAction.REJECT));
-    nat.setPoolIpFirst(new Ip("4.5.6.7"));
-
-    Flow transformed =
-        BdpEngine.applySourceNat(
-            flow, null, ImmutableMap.of(), ImmutableMap.of(), singletonList(nat));
-    assertThat(transformed, is(flow));
-  }
-
-  @Test
-  public void testApplySourceNatFirstMatchWins() {
-    Flow flow = makeFlow();
-
-    SourceNat nat = new SourceNat();
-    nat.setAcl(makeAcl("firstAccept", LineAction.ACCEPT));
-    nat.setPoolIpFirst(new Ip("4.5.6.7"));
-
-    SourceNat secondNat = new SourceNat();
-    secondNat.setAcl(makeAcl("secondAccept", LineAction.ACCEPT));
-    secondNat.setPoolIpFirst(new Ip("4.5.6.8"));
-
-    Flow transformed =
-        BdpEngine.applySourceNat(
-            flow, null, ImmutableMap.of(), ImmutableMap.of(), Lists.newArrayList(nat, secondNat));
-    assertThat(transformed.getSrcIp(), equalTo(new Ip("4.5.6.7")));
-  }
-
-  @Test
-  public void testApplySourceNatLateMatchWins() {
-    Flow flow = makeFlow();
-
-    SourceNat nat = new SourceNat();
-    nat.setAcl(makeAcl("rejectAll", LineAction.REJECT));
-    nat.setPoolIpFirst(new Ip("4.5.6.7"));
-
-    SourceNat secondNat = new SourceNat();
-    secondNat.setAcl(makeAcl("acceptAnyway", LineAction.ACCEPT));
-    secondNat.setPoolIpFirst(new Ip("4.5.6.8"));
-
-    Flow transformed =
-        BdpEngine.applySourceNat(
-            flow, null, ImmutableMap.of(), ImmutableMap.of(), Lists.newArrayList(nat, secondNat));
-    assertThat(transformed.getSrcIp(), equalTo(new Ip("4.5.6.8")));
-  }
-
-  @Test
-  public void testApplySourceNatInvalidAclThrows() {
-    Flow flow = makeFlow();
-
-    SourceNat nat = new SourceNat();
-    nat.setAcl(makeAcl("matchAll", LineAction.ACCEPT));
-
-    _thrown.expect(BatfishException.class);
-    _thrown.expectMessage("missing NAT address or pool");
-    BdpEngine.applySourceNat(flow, null, ImmutableMap.of(), ImmutableMap.of(), singletonList(nat));
   }
 
   private void testBgpAsPathMultipathHelper(
